+++ util/health.go
@@ -1,4 +1,4 @@
-package main
+package util
 
 import (
 	"context"
@@ -6,8 +6,6 @@ import (
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/health/grpc_health_v1"
 	"google.golang.org/grpc/status"
-
-	log "github.com/media-streaming-mesh/msm-dp/util"
 )
 
 // TODO - implement some checks to make sure all necessary resources/connections/interfaces are ready
@@ -21,20 +19,20 @@ func NewHealthChecker() *HealthChecker {
 }
 
 func (h *HealthChecker) Check(context.Context, *grpc_health_v1.HealthCheckRequest) (*grpc_health_v1.HealthCheckResponse, error) {
-	log.Infof("Serving the Check request for health check")
+	Infof("Serving the Check request for health check")
 
 	if isEverythingReady == true {
-		log.Debugf("✅ Server's status is %s", grpc_health_v1.HealthCheckResponse_SERVING)
+		Debugf("✅ Server's status is %s", grpc_health_v1.HealthCheckResponse_SERVING)
 		return &grpc_health_v1.HealthCheckResponse{
 			Status: grpc_health_v1.HealthCheckResponse_SERVING,
 		}, nil
 	} else if isEverythingReady == false {
-		log.Debugf("🚫 Server's status is %s", grpc_health_v1.HealthCheckResponse_NOT_SERVING)
+		Debugf("🚫 Server's status is %s", grpc_health_v1.HealthCheckResponse_NOT_SERVING)
 		return &grpc_health_v1.HealthCheckResponse{
 			Status: grpc_health_v1.HealthCheckResponse_NOT_SERVING,
 		}, nil
 	} else {
-		log.Debugf("🚫 Server's status is %s", grpc_health_v1.HealthCheckResponse_UNKNOWN)
+		Debugf("🚫 Server's status is %s", grpc_health_v1.HealthCheckResponse_UNKNOWN)
 		return &grpc_health_v1.HealthCheckResponse{
 			Status: grpc_health_v1.HealthCheckResponse_UNKNOWN,
 		}, nil
